+++ src/views/WebInputView.js
@@ -1,0 +1,12 @@
+export default class WebInputView {
+  static readPurchaseAmount() {
+    return document.getElementById('price')?.value;
+  }
+  static readWinNumbers() {
+    const winningNumberInputs = document.querySelectorAll('.winning-number');
+    return [...winningNumberInputs].map((winningNumberInput) => Number(winningNumberInput.value));
+  }
+  static readBonusNumber() {
+    return Number(document.querySelector('.bonus-number').value);
+  }
+}
