+++ src/views/index.js
@@ -1,2 +1,4 @@
 export { default as InputView } from './InputView.js';
 export { default as OutputView } from './OutputView.js';
+export { default as WebInputView } from './WebInputView.js';
+export { default as WebOutputView } from './WebOutputView.js';
